Move all codemodel subtypes outside
